+++ world-with-temperature/src/utils/MathUtils.h
@@ -6,10 +6,12 @@
 #define RECAST_MATHUTILS_H
 
 
+#include <cmath>
+
 namespace MathUtils {
 
     template <typename T>
-    inline T lerp(T a, T b, float t) {
+    inline T lerp(T a, T b, double t) {
         return (1 - t) * a + t * b;
     }
